Use a domain model for the log records
This is a quick and dirty solution which needs a lot of improvements.
The aim was to have a domain model without using extbase, as it's may
not ready yet when we log something and we also want to store log in
case extbase itself fails for some reason.

Suggestions for a better implementation of an extbase-free domain
model without relying on any third party packages are very welcome.